:muscle: Move `core/util/main.ts` to `util/main.ts`
