+++ denops/gin/util/main.ts
@@ -1,8 +1,8 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
-import { expand } from "../../util/expand.ts";
-import { findWorktreeFromDenops } from "../../git/worktree.ts";
+import { expand } from "./expand.ts";
+import { findWorktreeFromDenops } from "../git/worktree.ts";
 
 export function main(denops: Denops): void {
   denops.dispatcher = {
